+++ dialogflow/apiv2/entity_types_client.go
@@ -232,8 +232,8 @@ func (c *EntityTypesClient) Close() error {
 // setGoogleClientInfo sets the name and version of the application in
 // the `x-goog-api-client` header passed on each request. Intended for
 // use by Google-written clients.
-func (c *EntityTypesClient) setGoogleClientInfo(...string) {
-	c.internalClient.setGoogleClientInfo()
+func (c *EntityTypesClient) setGoogleClientInfo(keyval ...string) {
+	c.internalClient.setGoogleClientInfo(keyval...)
 }
 
 // Connection returns a connection to the API service.