+++ internal/gapicgen/generator/config.go
@@ -79,7 +79,7 @@ var microgenGapicConfigs = []*microgenConfig{
 		pkg:                   "language",
 		importPath:            "cloud.google.com/go/language/apiv1",
 		gRPCServiceConfigPath: "google/cloud/language/v1/language_grpc_service_config.json",
-		apiServiceConfigPath:  "google/cloud/language/language_v1.yaml",
+		apiServiceConfigPath:  "google/cloud/language/v1/language_v1.yaml",
 		releaseLevel:          "ga",
 	},
 	{
@@ -1082,6 +1082,15 @@ var microgenGapicConfigs = []*microgenConfig{
 		releaseLevel: "beta",
 	},
 	{
+		inputDirectoryPath:    "google/cloud/essentialcontacts/v1",
+		pkg:                   "essentialcontacts",
+		importPath:            "cloud.google.com/go/essentialcontacts/apiv1",
+		gRPCServiceConfigPath: "google/cloud/essentialcontacts/v1/essentialcontacts_v1_grpc_service_config.json",
+		apiServiceConfigPath:  "google/cloud/essentialcontacts/v1/essentialcontacts_v1.yaml",
+		// GA after 2021/06/18
+		releaseLevel: "beta",
+	},
+	{
 		inputDirectoryPath:    "google/api/serviceusage/v1",
 		pkg:                   "serviceusage",
 		importPath:            "cloud.google.com/go/serviceusage/apiv1",
@@ -1089,6 +1098,75 @@ var microgenGapicConfigs = []*microgenConfig{
 		apiServiceConfigPath:  "google/api/serviceusage/v1/serviceusage_v1.yaml",
 		// GA after 2021/06/18
 		releaseLevel: "beta",
+	},
+	{
+		inputDirectoryPath:    "google/cloud/shell/v1",
+		pkg:                   "shell",
+		importPath:            "cloud.google.com/go/shell/apiv1",
+		gRPCServiceConfigPath: "google/cloud/shell/v1/cloudshell_grpc_service_config.json",
+		apiServiceConfigPath:  "google/cloud/shell/v1/cloudshell_v1.yaml",
+		// GA after 2021/06/26
+		releaseLevel: "beta",
+	},
+	{
+		inputDirectoryPath:    "google/cloud/vpcaccess/v1",
+		pkg:                   "vpcaccess",
+		importPath:            "cloud.google.com/go/vpcaccess/apiv1",
+		gRPCServiceConfigPath: "google/cloud/vpcaccess/v1/vpcaccess_grpc_service_config.json",
+		apiServiceConfigPath:  "google/cloud/vpcaccess/v1/vpcaccess_v1.yaml",
+		// GA after 2021/06/26
+		releaseLevel: "beta",
+	},
+	{
+		inputDirectoryPath:    "google/cloud/privatecatalog/v1beta1",
+		pkg:                   "privatecatalog",
+		importPath:            "cloud.google.com/go/privatecatalog/apiv1beta1",
+		gRPCServiceConfigPath: "google/cloud/privatecatalog/v1beta1/cloudprivatecatalog_grpc_service_config.json",
+		apiServiceConfigPath:  "google/cloud/privatecatalog/v1beta1/cloudprivatecatalog_v1beta1.yaml",
+		releaseLevel:          "beta",
+	},
+	{
+		inputDirectoryPath:    "google/cloud/tpu/v1",
+		pkg:                   "tpu",
+		importPath:            "cloud.google.com/go/tpu/apiv1",
+		gRPCServiceConfigPath: "google/cloud/tpu/v1/tpu_grpc_service_config.json",
+		apiServiceConfigPath:  "google/cloud/tpu/v1/tpu_v1.yaml",
+		// GA after 2021/07/02
+		releaseLevel: "beta",
+	},
+	{
+		inputDirectoryPath:    "google/cloud/apigeeconnect/v1",
+		pkg:                   "apigeeconnect",
+		importPath:            "cloud.google.com/go/apigeeconnect/apiv1",
+		gRPCServiceConfigPath: "google/cloud/apigeeconnect/v1/connection_grpc_service_config.json",
+		apiServiceConfigPath:  "google/cloud/apigeeconnect/v1/apigeeconnect_v1.yaml",
+		// GA after 2021/07/03
+		releaseLevel: "beta",
+	},
+	{
+		inputDirectoryPath:    "google/cloud/lifesciences/v2beta",
+		pkg:                   "lifesciences",
+		importPath:            "cloud.google.com/go/lifesciences/apiv2beta",
+		gRPCServiceConfigPath: "google/cloud/lifesciences/v2beta/lifesciences_grpc_service_config.json",
+		apiServiceConfigPath:  "google/cloud/lifesciences/v2beta/lifesciences_v2beta.yaml",
+		releaseLevel:          "beta",
+	},
+	{
+		inputDirectoryPath:    "google/cloud/aiplatform/v1",
+		pkg:                   "aiplatform",
+		importPath:            "cloud.google.com/go/aiplatform/apiv1",
+		gRPCServiceConfigPath: "google/cloud/aiplatform/v1/aiplatform_grpc_service_config.json",
+		apiServiceConfigPath:  "google/cloud/aiplatform/v1/aiplatform_v1.yaml",
+		// GA after 2021/07/03
+		releaseLevel: "beta",
+	},
+	{
+		inputDirectoryPath:    "google/cloud/gkeconnect/gateway/v1beta1",
+		pkg:                   "gateway",
+		importPath:            "cloud.google.com/go/gkeconnect/gateway/apiv1beta1",
+		gRPCServiceConfigPath: "google/cloud/gkeconnect/gateway/v1beta1/connectgw_grpc_service_config.json",
+		apiServiceConfigPath:  "google/cloud/gkeconnect/gateway/v1beta1/connectgateway_v1beta1.yaml",
+		releaseLevel:          "beta",
 	},
 
 	// Non-Cloud APIs
@@ -1115,5 +1193,14 @@ var microgenGapicConfigs = []*microgenConfig{
 		gRPCServiceConfigPath: "google/area120/tables/v1alpha1/tables_grpc_service_config.json",
 		apiServiceConfigPath:  "google/area120/tables/v1alpha1/area120tables_v1alpha1.yaml",
 		releaseLevel:          "alpha",
+	},
+	{
+		inputDirectoryPath:    "google/cloud/gsuiteaddons/v1",
+		pkg:                   "gsuiteaddons",
+		importPath:            "cloud.google.com/go/gsuiteaddons/apiv1",
+		gRPCServiceConfigPath: "google/cloud/gsuiteaddons/v1/gsuiteaddons_grpc_service_config.json",
+		apiServiceConfigPath:  "google/cloud/gsuiteaddons/v1/gsuiteaddons_v1.yaml",
+		// GA after 2021/06/10
+		releaseLevel: "beta",
 	},
 }