+++ memcache/apiv1beta2/cloud_memcache_client.go
@@ -143,8 +143,8 @@ func (c *CloudMemcacheClient) Close() error {
 // setGoogleClientInfo sets the name and version of the application in
 // the `x-goog-api-client` header passed on each request. Intended for
 // use by Google-written clients.
-func (c *CloudMemcacheClient) setGoogleClientInfo(...string) {
-	c.internalClient.setGoogleClientInfo()
+func (c *CloudMemcacheClient) setGoogleClientInfo(keyval ...string) {
+	c.internalClient.setGoogleClientInfo(keyval...)
 }
 
 // Connection returns a connection to the API service.