+++ networkconnectivity/apiv1alpha1/hub_client.go
@@ -175,8 +175,8 @@ func (c *HubClient) Close() error {
 // setGoogleClientInfo sets the name and version of the application in
 // the `x-goog-api-client` header passed on each request. Intended for
 // use by Google-written clients.
-func (c *HubClient) setGoogleClientInfo(...string) {
-	c.internalClient.setGoogleClientInfo()
+func (c *HubClient) setGoogleClientInfo(keyval ...string) {
+	c.internalClient.setGoogleClientInfo(keyval...)
 }
 
 // Connection returns a connection to the API service.